- Merge case 1 - 4 into 1 to make it more readable - add some comment for steps clarity
