--- Soal1Case4.java
@@ -1,5 +1,0 @@
-public class Soal1Case4 {
-    public static void main(String[] args) {
-        System.out.println("Soal 1 Case 4");
-    }
-}
